Apply C++20 and best-practice fixes from PR review
Bug fix:
- ComputeError: use abs() on both operands of the denominator and guard
  against zero-denominator; was silently wrong for negative values

C++20:
- Replace M_PI (POSIX extension) with std::numbers::pi
- Add Data() accessor to Array1D/2D/3D enabling std::span construction

API correctness:
- GridPolicy& -> const GridPolicy& in ConstituentState, RadiationField,
  and RadiationFieldComponents constructors
- [[nodiscard]] on SlantOpticalDepth, Dot, ComputeError
- Units() returns const std::string& (was copying on every call)
- FillRandom takes scalars by value (was by const ref)
- Explicit static_cast<double> for SZA in DeltaEddington::Solve
- FillRandom(TridiagonalMatrix) uses distinct seeds per diagonal

Documentation:
- Solve() warns that the input matrix A is modified in place

Cleanup:
- Remove unused #include <functional> from linear_algebra.hpp
- Add #include <cmath> to linear_algebra.inl (for std::abs on floats)
- Add assert(!states.empty()) to Accumulate() placeholder

+++ include/tuvx/radiative_transfer/solvers/delta_eddington.inl
@@ -63,7 +63,7 @@ namespace tuvx
           [&](std::size_t lev) { return vertical_grid.edges_(lev, grid_col); });
 
       SphericalGeometry geometry;
-      geometry.SetParameters(solar_zenith_angles[col], altitude_edges);
+      geometry.SetParameters(static_cast<double>(solar_zenith_angles[col]), altitude_edges);
 
       const std::size_t albedo_column = (surface_albedo.mid_point_values_.Size2() == 1) ? 0 : col;
